move authorized fetch logic to its own hook
